--- examples/face_detector/ios/FaceDetector/Extensions/UIImageOrientationExt.swift
@@ -1,30 +1,0 @@
-// Copyright 2023 The MediaPipe Authors.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-import UIKit
-
-extension UIImage.Orientation {
-  static func from(deviceOrientation: UIDeviceOrientation) -> UIImage.Orientation {
-    switch deviceOrientation {
-      case .portrait:
-        return .up
-      case .landscapeLeft:
-        return .left
-      case .landscapeRight:
-        return .right
-      default:
-        return .up
-    }
-  }
-}
